Zed integration
This PR introduces an experimental mode for Gemini to allow it to be used to
power the Agent panel within Zed.

This integration allows Zed to provide an alternative UI for Gemini that is
tightly integrated with the editor, showing diffs and file changes in context,
and letting the user edit files as they go.

Zed takes responsibility for the life-cycle, and boots Gemini with
`--experimental-acp`, before talking to it over our experimental ACP protocol.

Based on our experience making this work with Gemini and Zed, we hope (in the
future) to be able to re-use for multiple agents and IDEs. It currently covers
just what we needed to make this work: initialization and authorization, and the
ability to stream assistant messages and tool calls (including permission
prompts) to the UI.

Co-Authored-By: Agus Zubiaga <agus@zed.dev>
Co-Authored-By: Ben Brandt <benjamin.j.brandt@gmail.com>
Co-authored-by: mkorwel <matt.korwel@gmail.com>